requirements.txt
fastapi
uvicorn
rembg==2.0.61
onnxruntime==1.23.1
python-multipart
Pillow
